Unstable partial refactor of node.go
